new_requirements.txt
adal==1.2.7
ansible==2.8.6
asn1crypto==1.5.1
azure-common==1.1.28
azure-nspkg==3.0.2
azure-storage-common==1.3.0
azure-storage-nspkg==3.1.0
bcrypt==4.3.0
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==2.1.1
circuitbreaker==2.1.3
click==8.1.8
cryptography==3.4.8
filelock==3.18.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
oci==2.139.0
oscrypto==1.3.0
packaging==24.2
paramiko==3.3.1
pip-tools==7.4.1
pipdeptree==2.26.0
pycparser==2.22
pycryptodomex==3.22.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyproject_hooks==1.2.0
pypsrp==0.5.0
pyspnego==0.1.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
snowflake-connector-python==3.0.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==1.26.20
